OSAC-902: Move generic authorization resource type to IDP
Originally the type was in Keycloak's types but it should
be in the general IDP library. Also removes the keycloak
scope type since it's unnecessary and adds a few needed
constants.
