+++ internal/idp/const.go
@@ -1,0 +1,24 @@
+/*
+Copyright (c) 2026 Red Hat Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the
+License. You may obtain a copy of the License at
+
+  http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an
+"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific
+language governing permissions and limitations under the License.
+*/
+
+package idp
+
+// Authorization scope constants define actions that can be performed on protected resources.
+// These are used with Keycloak Authorization Services to control fine-grained access to Projects.
+const (
+	// ScopeViewProject allows viewing project details and status
+	ScopeViewProject = "VIEW_PROJECT"
+
+	// ScopeManageProject allows updating project metadata, deleting project, and managing permissions
+	ScopeManageProject = "MANAGE_PROJECT"
+)
